Доработала слайдер + правки по замечаниям
