+++ index.html
@@ -114,7 +114,7 @@
           <input id="price" name="price" type="number" placeholder="5000" max="100000" required
             data-pristine-max-message="Максимальная цена за ночь — 100000"
             data-pristine-required-message="Обязательное поле">
-          <div class="ad-form__slider" id="slider"></div>
+          <div class="ad-form__slider"></div>
         </fieldset>
         <fieldset class="ad-form__element ad-form__element--time">
           <label class="ad-form__label" for="timein">Время заезда и выезда</label>
