+++ js/const.js
@@ -2,15 +2,6 @@ const CENTER_TOKIO_LAT = 35.681729;
 const CENTER_TOKIO_LNG = 139.753927;
 const MAX_PRICE = 100000;
 const MIN_PRICE = 0;
-const MAX_ROOMS = 100;
-const MIN_GUESTS = 0;
-const SLIDER_STEP = 1;
-const MAP_ZOOM = 13;
-const MAIN_PIN_ICON_SIZE = [52, 52];
-const MAIN_PIN_ICON_ANCHOR = [26, 52];
-const PIN_ICON_SIZE = [40, 40];
-const PIN_ICON_ANCHOR = [20, 40];
 
-export {CENTER_TOKIO_LAT, CENTER_TOKIO_LNG, MAX_PRICE, MIN_PRICE, MAX_ROOMS, MIN_GUESTS, SLIDER_STEP, MAP_ZOOM,
-  MAIN_PIN_ICON_SIZE, MAIN_PIN_ICON_ANCHOR, PIN_ICON_SIZE, PIN_ICON_ANCHOR};
+export {CENTER_TOKIO_LAT, CENTER_TOKIO_LNG, MAX_PRICE, MIN_PRICE};
