+++ js/form-validation.js
@@ -1,5 +1,8 @@
-import {CENTER_TOKIO_LAT, CENTER_TOKIO_LNG, MAX_PRICE, MAX_ROOMS, MIN_GUESTS} from './const.js';
+import {CENTER_TOKIO_LAT, CENTER_TOKIO_LNG, MAX_PRICE} from './const.js';
 import './price-slider.js';
+
+const MAX_ROOMS = 100;
+const MIN_GUESTS = 0;
 
 const minPrice = {
   bungalow: 0,
