+++ js/main.js
@@ -1,11 +1,3 @@
 import './form-validation.js';
-import {map} from './map.js';
-import {disabledPage, activatedPage} from './page-state.js';
+import './map.js';
 
-if (map) {
-  map.on('load', () => {
-    activatedPage();
-  });
-} else {
-  disabledPage();
-}
